filterTwoTables/requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
